Restructure project to use Lerna
This is in preparation for the migration of the project to the @chaff
scope, and the addition of the Drowsy server library.
