+++ packages/drowsy-client/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@chaff/drowsy-client",
+  "version": "4.1.4",
+  "description": "Client library for Drowsy",
+  "keywords": [
+    "api",
+    "api-client",
+    "api-integration",
+    "rest",
+    "rest-client",
+    "rest-integration",
+    "restful",
+    "restful-client",
+    "restful-integration"
+  ],
+  "author": "Iain Reid <iainreid382@gmail.com> (https://chaff.land)",
+  "homepage": "https://github.com/iainreid820/drowsy#readme",
+  "license": "MIT",
+  "main": "./lib/drowsy.js",
+  "types": "./lib/drowsy.d.ts",
+  "files": [
+    "lib"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/iainreid820/drowsy.git"
+  },
+  "scripts": {
+    "prepare": "mkdir -p lib && cp src/main/* lib/",
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/iainreid820/drowsy/issues"
+  }
+}
