+++ packages/drowsy/client.js
@@ -1,0 +1,1 @@
+module.exports = require("@chaff/drowsy-client");
