+++ packages/drowsy/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@chaff/drowsy",
+  "version": "4.1.4",
+  "description": "The lazy integration tool for RESTful interfaces",
+  "keywords": [
+    "api",
+    "api-client",
+    "api-integration",
+    "rest",
+    "rest-client",
+    "rest-integration",
+    "restful",
+    "restful-client",
+    "restful-integration"
+  ],
+  "author": "Iain Reid <iainreid382@gmail.com> (https://chaff.land)",
+  "homepage": "https://github.com/iainreid820/drowsy#readme",
+  "license": "MIT",
+  "files": [
+    "client.d.ts",
+    "client.js"
+  ],
+  "dependencies": {
+    "@chaff/drowsy-client": "^4.1.4"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/iainreid820/drowsy.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/iainreid820/drowsy/issues"
+  }
+}
